requirements.txt
google-generativeai==0.3.1
python-dotenv==1.0.0
pyTelegramBotAPI==4.12.0
requests==2.31.0
huggingface_hub==0.22.0 